@@ -2,19 +2,16 @@
   "name": "exmaple",
   "version": "1.0.0",
   "description": "",
-
   "main": "dist/exmaple.ssr.js",
   "browser": "dist/exmaple.esm.js",
   "module": "dist/exmaple.esm.js",
   "unpkg": "dist/exmaple.min.js",
   "types": "dist/types/src/entry.esm.d.ts",
-
   "files": [
     "dist/*",
     "src/**/*.vue"
   ],
   "sideEffects": false,
-
   "scripts": {
     "serve": "vue-cli-service serve dev/serve.ts",
     "prebuild": "rimraf ./dist",
@@ -24,10 +21,7 @@
     "build:unpkg": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format iife",
     "postbuild": "rimraf ./dist/types/dev ./dist/types/src/entry.d.ts"
   },
-
-  "dependencies": {
-  },
-
+  "dependencies": {},
   "devDependencies": {
     "@babel/core": "^7.14.6",
     "@babel/preset-env": "^7.14.7",
@@ -40,21 +34,23 @@
     "@vue/cli-plugin-babel": "^4.5.13",
     "@vue/cli-plugin-typescript": "^4.5.13",
     "@vue/cli-service": "^4.5.13",
+    "@vue/compiler-sfc": "^3.0.11",
     "@zerollup/ts-transform-paths": "^1.7.18",
     "cross-env": "^7.0.3",
     "minimist": "^1.2.5",
+    "postcss": "^8.2.10",
     "rimraf": "^3.0.2",
     "rollup": "^2.52.8",
+    "rollup-plugin-postcss": "^4.0.0",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-vue": "^5.1.9",
+    "rollup-plugin-vue": "^6.0.0",
     "rollup-plugin-typescript2": "^0.30.0",
     "ttypescript": "^1.5.12",
     "typescript": "^4.0.3",
-    "vue": "^2.6.14",
-    "vue-template-compiler": "^2.6.14"
+    "vue": "^3.0.5"
   },
   "peerDependencies": {
-    "vue": "^2.6.14"
+    "vue": "^3.0.5"
   },
   "engines": {
     "node": ">=12"
