@@ -1,9 +1,9 @@
 <script lang="ts">
-import Vue from 'vue';
+import { defineComponent } from 'vue';
 // Uncomment import and local "components" registration if library is not registered globally.
 // import { ExmapleSample } from '@/entry.esm';
 
-export default Vue.extend({
+export default defineComponent({
   name: 'ServeDev',
   // components: {
   //  ExmapleSample,
