@@ -1,2 +1,4 @@
 /* eslint-disable import/prefer-default-export */
-export { default as ExmapleSample } from './exmaple-sample.vue';
+import ExampleSample from './exmaple-sample.vue';
+
+export { ExampleSample };
